Move to the very beginning before the 1st time using the pref
